move footer files into footer-dir folder
