+++ src/app/app.module.ts
@@ -6,9 +6,9 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MaterialModule } from './material/material.module';
-import { FooterComponent } from './footer/footer.component';
+import { FooterComponent } from './footer-dir/footer/footer.component';
 import { HeaderComponent } from './header/header.component';
-import { AboutComponent } from './about/about.component';
+import { AboutComponent } from './footer-dir/about/about.component';
 import { ContentsComponent } from './contents/contents.component';
 import { HomeComponent } from './home/home.component';
 import { CardsComponent } from './activities/cards/cards.component';
@@ -21,23 +21,23 @@ import { ExpandablesComponent } from './activities/expandables/expandables.compo
 import { DuelsComponent } from './activities/duels/duels.component';
 import { ConvertTimePipe } from './convert-time.pipe';
 import { TimedCardsComponent } from './activities/timed-cards/timed-cards.component';
-import { ChangelogComponent } from './changelog/changelog.component';
-import { TeachersGuideComponent } from './teachers-guide/teachers-guide.component';
-import { RoadmapComponent } from './roadmap/roadmap.component';
-import { AbbreviationsComponent } from './abbreviations/abbreviations.component';
+import { ChangelogComponent } from './footer-dir/changelog/changelog.component';
+import { TeachersGuideComponent } from './footer-dir/teachers-guide/teachers-guide.component';
+import { RoadmapComponent } from './footer-dir/roadmap/roadmap.component';
+import { AbbreviationsComponent } from './footer-dir/abbreviations/abbreviations.component';
 import { EscapeRoomComponent } from './activities/escape-room/escape-room.component';
 import { ContextComponent } from './activities/context/context.component';
 import { BoardGameComponent } from './activities/board-game/board-game.component';
 import { DialogBoardGameComponent } from './activities/dialog-board-game/dialog-board-game.component';
 import { RankingComponent } from './activities/ranking/ranking.component';
 import { StoryPuzzleComponent } from './activities/story-puzzle/story-puzzle.component';
-import { DictionariesComponent } from './dictionaries/dictionaries.component';
-import { CorporaComponent } from './corpora/corpora.component';
-import { MiscComponent } from './misc/misc.component';
+import { DictionariesComponent } from './footer-dir/dictionaries/dictionaries.component';
+import { CorporaComponent } from './footer-dir/corpora/corpora.component';
+import { MiscComponent } from './footer-dir/misc/misc.component';
 import { SentenceViewComponent } from './activities/story-puzzle/sentence-view/sentence-view.component';
-import { NewsletterSignupComponent } from './newsletter-signup/newsletter-signup.component';
-import { TextComponent } from './text/text.component';
-import { VideoComponent } from './video/video.component';
+import { NewsletterSignupComponent } from './footer-dir/newsletter-signup/newsletter-signup.component';
+import { TextComponent } from './footer-dir/text/text.component';
+import { VideoComponent } from './footer-dir/video/video.component';
 
 @NgModule({
   declarations: [
