+++ src/app/footer-dir/abbreviations/abbreviations.component.spec.ts
