+++ src/app/footer-dir/abbreviations/abbreviations.component.ts
