+++ src/app/footer-dir/about/about.component.ts
