+++ src/app/footer-dir/corpora/corpora.component.css
