+++ src/app/footer-dir/dictionaries/dictionaries.component.spec.ts
