+++ src/app/footer-dir/dictionaries/dictionaries.component.ts
