+++ src/app/footer-dir/footer/footer.component.spec.ts
