+++ src/app/footer-dir/misc/misc.component.ts
