+++ src/app/footer-dir/newsletter-signup/newsletter-signup.component.spec.ts
