+++ src/app/footer-dir/teachers-guide/teachers-guide.component.spec.ts
