+++ src/app/footer-dir/teachers-guide/teachers-guide.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { Utils } from '../utils';
+import { Utils } from '../../utils';
 
 @Component({
   selector: 'app-teachers-guide',
