+++ src/app/footer-dir/text/text.component.html
