+++ src/app/footer-dir/text/text.component.spec.ts
