+++ src/app/footer-dir/video/video.component.ts
